Moved snackbar content in separate component
